Add check if post exists in like and unlike actions of PostViewSet Use specific exception instead of abstract Follow PEP8

+++ social_network/views.py
@@ -1,13 +1,12 @@
-from django.shortcuts import render
-from rest_framework import viewsets
-from social_network.models import Post,Like
+from rest_framework import viewsets, status
+from social_network.models import Post, Like
 from social_network.serializers import PostSerializer, LikeSerializer, UserSerializer
 from django.contrib.auth.models import User
 from rest_framework.decorators import action
 from rest_framework import permissions
-from django.utils import timezone
 from rest_framework.response import Response
 from social_network.permissions import UserPermission
+from django.db.utils import IntegrityError
 
 
 class PostViewSet(viewsets.ModelViewSet):
@@ -17,33 +16,36 @@ class PostViewSet(viewsets.ModelViewSet):
 
     @action(detail=True, methods=['post'])
     def like(self, request, pk=None):
-        post = Post.objects.get(id=pk)
+        post = Post.objects.filter(id=pk).first()
+        if not post:
+            return Response({'status': 'Post does not exist.'}, status=status.HTTP_400_BAD_REQUEST)
         user = request.user
         try:
-            old_like = Like.objects.get(post=post,liker=user)
-            return Response({'response': "post is already liked before"})
-        except:
-            new_like = Like(post = post, liker = user, like_date = timezone.now())
-            new_like.save()
-            return Response({'response' : "successfully liked"})
+            like = Like(post=post, liker=user)
+            like.save()
+            return Response({'status': "post is successfully liked"})
+        except IntegrityError:
+            return Response({'status': "post is already liked"})
 
-    @action(detail=True, methods = ['post'])
+    @action(detail=True, methods=['post'])
     def unlike(self, request, pk=None):
-        post = Post.objects.get(id=pk)
+        post = Post.objects.filter(id=pk).first()
+        if not post:
+            return Response({'status': 'Post does not exist.'}, status=status.HTTP_400_BAD_REQUEST)
         user = request.user
         try:
             old_like = Like.objects.get(post=post, liker=user)
             old_like.delete()
             return Response({'response': "post is successfully unliked"})
-        except:
+        except Like.DoesNotExist:
             return Response({'response': "post was not liked before"})
-
 
 
 class LikeViewSet(viewsets.ModelViewSet):
     queryset = Like.objects.all()
     serializer_class = LikeSerializer
     permission_classes = [permissions.IsAuthenticated]
+
 
 class UserViewSet(viewsets.ModelViewSet):
     queryset = User.objects.all()
